+++ sharding-jdbc-orchestration-spring/sharding-jdbc-orchestration-spring-namespace/src/main/java/io/shardingjdbc/orchestration/spring/namespace/constants/ShardingDataSourceBeanDefinitionParserTag.java
@@ -67,4 +67,6 @@ public final class ShardingDataSourceBeanDefinitionParserTag {
     public static final String COLUMN_KEY_GENERATOR_CLASS = "column-key-generator-class";
     
     public static final String KEY_GENERATOR_CLASS = "key-generator-class";
+    
+    public static final String LOGIC_INDEX_ATTRIBUTE = "logic-index";
 }